엔진 코드 기능 추가 및 정리. - 이미지 반전 기능 추가.

+++ Engine/Math/Vector2.cpp
@@ -2,7 +2,7 @@
 #include "Vector2.h"
 #include <cassert>
 
-Vector2::Vector2(int x, int y)
+Vector2::Vector2(float x, float y)
     : x(x), y(y)
 {
 }
